==> 1_Pulling from pubmed.ipynb <==
import pandas as pd
import urllib
import requests
import re
import glob
import json

This notebook contains the original data collection code. The files used to begin this work are located at https://www.ncbi.nlm.nih.gov/pmc/tools/ftp/#index_files

Using those files, we then ping the PMC api to get the paper structures.

The order of cells in this notebook is not accurate, since I ended up needing to restart and rerun various bits over the course of data collection. I also can't really verify that everything is still here.


# Get the pubmed papers

#bioc-pmc files
files_to_get = pd.read_table('/home/riddleta/ac_knowl/data/filelist.txt', sep='\s+')
files_to_get.shape

oa_comm_use_files = pd.read_csv('/home/riddleta/ac_knowl/data/oa_comm_use_file_list.txt', 
                                sep='\\t', engine='python', header=None, skiprows=1)
oa_comm_use_files.columns = ['File', 'ref', 'PMCID', 'PMID', 'license']
oa_comm_use_files.shape

oa_files = pd.read_table('/home/riddleta/ac_knowl/data/oa_file_list.txt', 
                         sep='\\t', engine='python', header=None, skiprows=1)
oa_files.columns = ['File', 'ref', 'PMCID', 'PMID', 'license']
oa_files.shape

unique_ids = pd.concat([oa_files['PMCID'], oa_comm_use_files['PMCID'], files_to_get['PMCID']]).unique()

## Main function to return data from fulltext search

def return_useful_data(r):
    git_hits = 0
    osf_hits = 0
    nda_hits = 0
    open_neuro = 0
    fmri = 0
    res = 0
    r_out = r.json()
    date = r_out['date']
    src = r_out['source']
    try:
        doi = r_out['documents'][0]['passages'][0]['infons']['article-id_doi']
    except:
        doi = None
    try:
        yr = r_out['documents'][0]['passages'][0]['infons']['year']
    except:
        yr = None
    if yr not in ['2008', '2009', '2010', '2011', '2012', '2013', 
                  '2014', '2015', '2016', '2017', '2018', '2019']:
        return(date, src, doi, git_hits, osf_hits, nda_hits, open_neuro, fmri, res, yr)
    else:
        for passage in r_out['documents'][0]['passages']:
            if re.search(re.compile('github'), passage['text'].lower()):
                git_hits = git_hits + 1
            if re.search(re.compile('osf\.io'), passage['text'].lower()):
                osf_hits = osf_hits + 1
            if re.search(re.compile('nda\.nih\.gov'), passage['text'].lower()):
                nda_hits = nda_hits + 1
            if re.search(re.compile('openneuro'), passage['text'].lower()):
                open_neuro = open_neuro + 1
            if re.search(re.compile('fmri'), passage['text'].lower()):
                fmri = fmri + 1
            if re.search(re.compile('results'), passage['text'].lower()):
                res = res + 1
        
        return(date, src, doi, git_hits, osf_hits, nda_hits, open_neuro, fmri, res, yr)

## loop to get everything
realistically, this took about two weeks. I ran this in small batches and saved the files iteratively. The second code block gives an example. There ended up being 168 papers that we couldn't retrieve.

#git_links = []
#missed = []

#for i, j in enumerate(unique_ids):
#    pth = 'https://www.ncbi.nlm.nih.gov/research/bionlp/RESTful/pmcoa.cgi/BioC_json/' + j + '/unicode'
#    try:
#        r = requests.get(pth)
#        git_links.append(return_useful_data(r) + (j, i))
#    except:
#        missed.append([i,j])

#df = pd.DataFrame(git_links, columns=['date', 'src', 'doi', 'git_hits', 
#                                      'osf_hits', 'nda_hits', 'open_neuro', 
#                                      'fmri', 'res', 'yr', 'pmcid', 'idx'])
#df.to_csv('/home/riddleta/ac_knowl/output/bionlp_03.csv')

# like this:

#git_links = []
#missed = []
#progress = [i for i in range(0, 15000, 2500)]

#for i, j in enumerate(unique_ids[625000:]):
#    if i in progress:
#        print(i)
#    pth = 'https://www.ncbi.nlm.nih.gov/research/bionlp/RESTful/pmcoa.cgi/BioC_json/' + j + '/unicode'
#    try:
#        r = requests.get(pth)
#        git_links.append(return_useful_data(r) + (j, i))
#    except:
#        missed.append([i,j])
#    if i==15000:
#        break

## target the subset of papers I want for a full text download

files = glob.glob('/home/riddleta/ac_knowl/output/bio*.csv')
file_list = []
for filename in files:
    df = pd.read_csv(filename, index_col=None, header=0)
    file_list.append(df)

frame = pd.concat(file_list, axis=0, ignore_index=True)

### Some of the years were formatted incorrectly or missed entirely. This resolves that.

frame['yr_fixed'] = frame['yr'].replace({'2016;':'2016'})

pmids = frame[pd.isna(frame.yr_fixed)].pmcid.tolist()
len(pmids)

for pmid in pmids:
    pth = 'https://www.ncbi.nlm.nih.gov/research/bionlp/RESTful/pmcoa.cgi/BioC_json/' + pmid + '/unicode'
    r = requests.get(pth)
    r_out = r.json()
    for i in range(0, 5):
        if 'year' in r_out['documents'][0]['passages'][i]['infons'].keys():
            year = r_out['documents'][0]['passages'][i]['infons']['year']
    
    frame['yr_fixed'][frame.pmcid==pmid] = year
            

frame['yr_ints'] = frame.yr_fixed.astype(int)

df_targets = frame[frame.yr_ints>2008]

df_targets.shape

def already_retrieved_papes(dirs):
    pmcids = []
    for d in dirs:

        with open(d) as infile:
            dat = json.load(infile)
            for paper in dat:
                pmcids.append(paper['documents'][0]['id'])
    
    return(['PMC'+i for i in pmcids])

def missed_papes(dirs):
    pmcids = []
    for d in dirs:
        with open(d, 'r') as infile:
            dat = infile.readlines()
            for paper in dat:
                pmcids.append(paper[0:10])
    return(pmcids)

already_retrieved = glob.glob('/home/riddleta/ac_knowl/output/full_texts/papes*')

gots = already_retrieved_papes(already_retrieved)

missed = glob.glob('/home/riddleta/ac_knowl/output/full_texts/misses*')

misses = missed_papes(missed)
len(misses)

df_targets_sub = df_targets[~df_targets.pmcid.isin(gots)]
df_targets_sub.shape

def already_retrieved_papes(dirs):
    pmcids = []
    for d in dirs:

        with open(d) as infile:
            dat = json.load(infile)
            for paper in dat:
                pmcids.append(paper['documents'][0]['id'])
    
    return(['PMC'+i for i in pmcids])

already_retrieved = glob.glob('/home/riddleta/ac_knowl/output/full_texts/papes*')

gots = already_retrieved_papes(already_retrieved)

df_targets_sub = df_targets[~df_targets.pmcid.isin(gots)]
df_targets_sub.shape

file_writes = [i for i in range(1827500, 2700000, 2500)]


outdat = []
missed = []
for i, pmcid in enumerate(df_targets_sub.pmcid):
    try:
        pth = 'https://www.ncbi.nlm.nih.gov/research/bionlp/RESTful/pmcoa.cgi/BioC_json/' + pmcid + '/unicode'
        r = requests.get(pth)
        r_out = r.json()
        outdat.append(r_out)
    except:
        missed.append(pmcid)
    if i+1827500 in file_writes: 
        with open('/home/riddleta/ac_knowl/output/full_texts/papes_'+str(i+1827500)+'.txt', 'w') as fout:
            json.dump(outdat, fout)
        if len(missed)>0:
            with open('/home/riddleta/ac_knowl/output/full_texts/misses_'+str(i+1827500)+'.txt', 'w') as f:
                for item in missed:
                    f.write("%s\n" % item)
        outdat = []
        missed = []


with open('/home/riddleta/ac_knowl/output/full_texts/papes_7500.txt') as infile:
    dat = json.load(infile)


outdat[0]

pmid = df_targets.pmcid.iloc[0]

pth = 'https://www.ncbi.nlm.nih.gov/research/bionlp/RESTful/pmcoa.cgi/BioC_json/' + pmid + '/unicode'
r = requests.get(pth)
r_out = r.json()


pmid

#r_out['documents'][0]['passages'][0]['infons']['year']
r_out['documents'][0]['id']# that's the pmcid

yr_counts = frame.yr_fixed.value_counts()
yr_counts.iloc[25:50]
