==> 2_Creating_index_file.ipynb <==
import pandas as pd
import glob
import json
import requests
import numpy as np

# Creating an index file
One of the problems with this database as it stands is that I'm not sure where to find the fulltext of a given paper.

To solve this, I'm going to create an index file that has the filepath and paper number for each paper we might be interested in. This will serve as a sort of master list of where to find things.

## Get locations of everything
Note that somehow I got duplicates of some papers, so this loop goes through and keeps all the originals, and puts the duplicates aside

already_retrieved = glob.glob('/home/riddleta/ac_knowl/output/full_texts/papes*')

status_prints = [i for i in range(0, 3000000, 100000)]

file_string = []
pmcid = []
paper_number = []
pmid = []

file_string_dupes = []
pmcid_dupes = []
paper_number_dupes = []
pmid_dupes = []

for i in already_retrieved:
    with open(i) as infile:
        dat = json.load(infile)
        for j, paper in enumerate(dat):
            if len(pmcid)+len(pmcid_dupes) in status_prints:
                print(len(pmcid))
            pap_id = paper['documents'][0]['id']
            if pap_id not in pmcid:
                try:
                    pmid.append(paper['documents'][0]['passages'][0]['infons']['article-id_pmid'])
                except:
                    pmid.append(np.nan)
                pmcid.append(pap_id)
                file_string.append(i)
                paper_number.append(j)
            else:
                pmcid_dupes.append(pap_id)
                file_string_dupes.append(i)
                paper_number_dupes.append(j)
                try:
                    pmid_dupes.append(paper['documents'][0]['passages'][0]['infons']['article-id_pmid'])
                except:
                    pmid_dupes.append(np.nan)

df = pd.DataFrame({'file':file_string,
                   'pmcid':pmcid,
                   'paper_number':paper_number})

df_dupes = pd.DataFrame({'file':file_string_dupes,
                        'pmcid':pmcid_dupes,
                        'paper_number':paper_number_dupes})

df.to_csv('/home/riddleta/ac_knowl/output/file_index.csv', index=False)

df_dupes.to_csv('/home/riddleta/ac_knowl/output/file_index_dupes.csv', index=False)

df.shape

# Notes


for grants:
https://dpcpsi.nih.gov/eo/intranet/opa/training/class_material

and NIH reporter

and original icite paper

other metadata: https://www.ncbi.nlm.nih.gov/pmc/tools/get-metadata/

already_retrieved = glob.glob('/home/riddleta/ac_knowl/output/full_texts/papes*')

with open(already_retrieved[0]) as infile:
    dat = json.load(infile)

for paper in dat:
    header = paper['documents'][0]['passages'][0]['infons']
    try:
        print(header['article-id_pmid'] + '*******' + header['article-id_pmc'])
    except:
        print(header['article-id_pmc'])

header
