==> 2a-FALSESTART_pull_paper_metadata.ipynb <==
import glob
import pandas as pd
import urllib
import requests
import xml.etree.ElementTree as ET
import time
import numpy as np
from itertools import zip_longest
import re

# False start

The code in this notebook is around an attempt to pull out some meta-data about these papers from some other APIs and from the papers themselves. Largely this was an attempt to get grant information. I eventually realized that much of the information I needed was in federal reporter so this code is not needed at the moment.

def grouper(iterable, n, fillvalue=None):
    args = [iter(iterable)] * n
    return zip_longest(*args, fillvalue=fillvalue)

df = pd.read_csv('/home/riddleta/ac_knowl/output/file_index.csv')

df.head()

with open('api_key.txt', 'r') as infile:
    line = infile.readlines()
    api_key = line[0][10:]

mesh_labs = []
mesh_ids = []
grant_ids = []
grant_agency = []
journal_out = []
pmid_out = []
auths_out = []
pmcid_out = []
total_papes = 0
for i in grouper(df.pmcid.tolist(), 200):
    try:
        i = list(map(str, i)) #ensure strings
        pth = 'https://eutils.ncbi.nlm.nih.gov/entrez/eutils/esummary.fcgi?db=pmc&id=' + str(','.join(i)) + '&retmode=json&api_key=' + str(api_key)
        r = requests.get(pth)
        r_out = r.json()
        jr=[] #store local group results
        pmr=[]
        ar=[]
        pmcr=[]
        for j in r_out['result']['uids']:
            jr.append(r_out['result'][str(j)]['fulljournalname'])
            pmr.append([art_id['value'] for art_id in r_out['result'][str(j)]['articleids'] if art_id['idtype']=='pmid'][0])
            ar.append(r_out['result'][str(j)]['authors'])
            pmcr.append(j)


        journal_out.extend(jr) #build up lists for writing
        pmid_out.extend(pmr)
        auths_out.extend(ar)
        pmcid_out.extend(pmcr)

        out_df = pd.DataFrame({'journal':journal_out,
                          'pmid':pmid_out,
                          'auths':auths_out,
                          'pmcid':pmcid_out})
        if out_df.shape[0]>100000:
            total_papes = total_papes + out_df.shape[0]
            fn = 'grant_id_'
            out_df.to_csv('/home/riddleta/ac_knowl/output/grant_identifiers/'+fn+str(total_papes)+'.csv', index=False)
            journal_out = []
            pmid_out = []
            auths_out = []
            pmcid_out = []

        time.sleep(.11)
    except Exception as e:
        print(str(i[0])+'*******'+str(i[199]))
        print(e)
        time.sleep(.11)


total_papes = total_papes + out_df.shape[0]
fn = 'grant_id_'
out_df.to_csv('/home/riddleta/ac_knowl/output/grant_identifiers/'+fn+str(total_papes)+'.csv', index=False)

total_papes

out_df.shape

li = []
all_files = glob.glob('/home/riddleta/ac_knowl/output/grant_identifiers/*')
for filename in all_files:
    tempdat = pd.read_csv(filename, index_col=None, header=0)
    li.append(tempdat)

frame = pd.concat(li, axis=0, ignore_index=True)

frame.shape

stragglers = df[~df.pmcid.isin(frame.pmcid)]
stragglers.shape

mesh_labs = []
mesh_ids = []
grant_ids = []
grant_agency = []
journal_out = []
pmid_out = []
auths_out = []
pmcid_out = []
#total_papes = 2504999
for i in grouper(stragglers.pmcid.tolist(), 200):
    try:
        i = list(map(str, i)) #ensure strings
        pth = 'https://eutils.ncbi.nlm.nih.gov/entrez/eutils/esummary.fcgi?db=pmc&id=' + str(','.join(i)) + '&retmode=json&api_key=' + str(api_key)
        r = requests.get(pth)
        r_out = r.json()
        jr=[] #store local group results
        pmr=[]
        ar=[]
        pmcr=[]
        for j in r_out['result']['uids']:
            try:
                jr.append(r_out['result'][str(j)]['fulljournalname'])
                pmr.append([art_id['value'] for art_id in r_out['result'][str(j)]['articleids'] if art_id['idtype']=='pmid'][0])
                ar.append(r_out['result'][str(j)]['authors'])
                pmcr.append(j)
            except:
                jr.append(np.nan)
                pmr.append(np.nan)
                ar.append(np.nan)
                pmcr.append(j)


        journal_out.extend(jr) #build up lists for writing
        pmid_out.extend(pmr)
        auths_out.extend(ar)
        pmcid_out.extend(pmcr)

        out_df = pd.DataFrame({'journal':journal_out,
                          'pmid':pmid_out,
                          'auths':auths_out,
                          'pmcid':pmcid_out})
        if out_df.shape[0]>100000:
            total_papes = total_papes + out_df.shape[0]
            fn = 'grant_id_'
            out_df.to_csv('/home/riddleta/ac_knowl/output/grant_identifiers/'+fn+str(total_papes)+'.csv', index=False)
            journal_out = []
            pmid_out = []
            auths_out = []
            pmcid_out = []

        time.sleep(.11)
    except Exception as e:
        print(str(i[0])+'*******'+str(i[199]))
        print(e)
        time.sleep(.11)

total_papes = total_papes + out_df.shape[0]
fn = 'grant_id_'
out_df.to_csv('/home/riddleta/ac_knowl/output/grant_identifiers/'+fn+str(total_papes)+'.csv', index=False)

li = []
all_files = glob.glob('/home/riddleta/ac_knowl/output/grant_identifiers/*')
for filename in all_files:
    tempdat = pd.read_csv(filename, index_col=None, header=0)
    li.append(tempdat)

frame = pd.concat(li, axis=0, ignore_index=True)
frame = frame.drop_duplicates()
frame['pmcid'] = frame.pmcid.astype('str')

frame.shape

df['pmcid'] = df.pmcid.astype('str')
df.shape

stragglers = df[~df.pmcid.isin(frame.pmcid)]
stragglers.shape

journal_out = []
pmid_out = []
auths_out = []
pmcid_out = []
for i in stragglers.pmcid.tolist():
    try:
        pth = 'https://eutils.ncbi.nlm.nih.gov/entrez/eutils/esummary.fcgi?db=pmc&id=' + str(i) + '&retmode=json&api_key=' + str(api_key)
        r = requests.get(pth)
        r_out = r.json()
        journal_out.append(r_out['result'][str(i)]['fulljournalname'])
        pmid_out.append([art_id['value'] for art_id in r_out['result'][str(i)]['articleids'] if art_id['idtype']=='pmid'][0])
        auths_out.append(r_out['result'][str(i)]['authors'])
        pmcid_out.append(i)

        time.sleep(.11)
    except Exception as e:
        print(str(i))
        print(e)
        journal_out.append(np.nan)
        pmid_out.append(np.nan)
        auths_out.append(np.nan)
        pmcid_out.append(i)
        time.sleep(.11)


out_df = pd.DataFrame({'journal':journal_out,
                  'pmid':pmid_out,
                  'auths':auths_out,
                  'pmcid':pmcid_out})

total_papes = total_papes + out_df.shape[0]
fn = 'grant_id_'
out_df.to_csv('/home/riddleta/ac_knowl/output/grant_identifiers/'+fn+str(total_papes)+'.csv', index=False)

## Hit the grant API
pull the mesh labels, ids, grant ids, and grant agency for each paper

li = []
all_files = glob.glob('/home/riddleta/ac_knowl/output/pmid_pmcid_link/*')
for filename in all_files:
    tempdat = pd.read_csv(filename, index_col=None, header=0)
    li.append(tempdat)

frame = pd.concat(li, axis=0, ignore_index=True)
frame['pmcid'] = frame.pmcid.astype('str')
#frame['pmid'] = frame.pmid.astype('int')
#frame['pmid'] = frame.pmid.astype('str')
frame = frame.drop_duplicates()


frame['pmid2'] = frame.pmid.fillna(0).astype('int')
frame.pmid2.head()

frame.shape

#initialize some lists to hold data at the paper level
paper_level_meshlabs = []
paper_level_meshids = []
paper_level_grant_ids = []
paper_level_grant_agency = []
paper_level_kw = []
pmid_out = []
#keep track of how many papers
total_papes = 0
fn = 'grant_id_'
#for each group of 200 papers
for i in grouper(frame.pmid2.tolist(), 200):
    try:
        i = list(map(str, i)) #ensure strings
        pth = 'https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi?db=pubmed&id=' + str(','.join(i)) + '&retmode=xml&api_key=' + str(api_key)
        #get the data
        r = requests.get(pth)
        root = ET.fromstring(r.content)
        # for each paper in the group of 200
        for paper in root.iter('PubmedArticle'):
            #initialize some lists to hold the variable-length content from a paper
            tmp_mesh_labs = []
            tmp_mesh_ids = []
            tmp_grant_ids = []
            tmp_grant_agency = []
            tmp_kw = []
            if paper.find('.//MeshHeading'):
                for mesh in paper.iter('MeshHeading'):
                    tmp_mesh_labs.append([j.text for j in mesh])
                    tmp_mesh_ids.append([j.attrib['UI'] for j in mesh])
            else:
                tmp_mesh_labs.append(np.nan)
                tmp_mesh_ids.append(np.nan)
            if paper.find('.//KeywordList'):
                for keyword in paper.iter('KeywordList'):
                    tmp_kw.append([j.text for j in keyword])
            else:
                tmp_kw.append(np.nan)
            if paper.find('.//Grant'):
                for grant in paper.iter('Grant'):
                    tmp_grant_ids.append([j.text for j in grant if j.tag=='GrantID'])
                    tmp_grant_agency.append([j.text for j in grant if j.tag=='Agency'])
            else:
                tmp_grant_ids.append(np.nan)
                tmp_grant_agency.append(np.nan)

            paper_level_meshlabs.append(tmp_mesh_labs)
            paper_level_meshids.append(tmp_mesh_ids)
            paper_level_grant_ids.append(tmp_grant_ids)
            paper_level_grant_agency.append(tmp_grant_agency)
            paper_level_kw.append(tmp_kw)
            pmid_out.append(paper[0][0].text)
    
        out_df = pd.DataFrame({'mesh':paper_level_meshlabs,
                               'mesh_id':paper_level_meshids,
                               'grant_ids':paper_level_grant_ids,
                               'grant_agency':paper_level_grant_agency,
                               'keyword':paper_level_kw,
                               'pmid':pmid_out})
        
        if out_df.shape[0]>100000:
            total_papes = total_papes + out_df.shape[0]
            out_df.to_csv('/home/riddleta/ac_knowl/output/grant_identifiers/'+fn+str(total_papes)+'.csv', index=False)
            paper_level_meshlabs = []
            paper_level_meshids = []
            paper_level_grant_ids = []
            paper_level_grant_agency = []
            paper_level_kw = []
            pmid_out = []
        
        time.sleep(.11)
            
    except Exception as e:
        print(str(i[0])+'*******'+str(i[199]))
        print(e)
        time.sleep(.11)

total_papes

total_papes = total_papes + out_df.shape[0]
out_df.to_csv('/home/riddleta/ac_knowl/output/grant_identifiers/'+fn+str(total_papes)+'.csv', index=False)

li = []
all_files = glob.glob('/home/riddleta/ac_knowl/output/grant_identifiers/*')
for filename in all_files:
    tempdat = pd.read_csv(filename, index_col=None, header=0)
    li.append(tempdat)

done_dat = pd.concat(li, axis=0, ignore_index=True)

stragglers = frame[~frame.pmid2.isin(done_dat.pmid)]
stragglers.shape

#initialize some lists to hold data at the paper level
paper_level_meshlabs = []
paper_level_meshids = []
paper_level_grant_ids = []
paper_level_grant_agency = []
paper_level_kw = []
pmid_out = []
#keep track of how many papers
total_papes = 0
fn = 'grant_id_'
#for each group of 200 papers
for i in grouper(stragglers.pmid2.tolist(), 200):
    try:
        i = list(map(str, i)) #ensure strings
        pth = 'https://eutils.ncbi.nlm.nih.gov/entrez/eutils/efetch.fcgi?db=pubmed&id=' + str(','.join(i)) + '&retmode=xml&api_key=' + str(api_key)
        #get the data
        r = requests.get(pth)
        root = ET.fromstring(r.content)
        # for each paper in the group of 200
        for paper in root.iter('PubmedArticle'):
            #initialize some lists to hold the variable-length content from a paper
            tmp_mesh_labs = []
            tmp_mesh_ids = []
            tmp_grant_ids = []
            tmp_grant_agency = []
            tmp_kw = []
            if paper.find('.//MeshHeading'):
                for mesh in paper.iter('MeshHeading'):
                    tmp_mesh_labs.append([j.text for j in mesh])
                    tmp_mesh_ids.append([j.attrib['UI'] for j in mesh])
            else:
                tmp_mesh_labs.append(np.nan)
                tmp_mesh_ids.append(np.nan)
            if paper.find('.//KeywordList'):
                for keyword in paper.iter('KeywordList'):
                    tmp_kw.append([j.text for j in keyword])
            else:
                tmp_kw.append(np.nan)
            if paper.find('.//Grant'):
                for grant in paper.iter('Grant'):
                    tmp_grant_ids.append([j.text for j in grant if j.tag=='GrantID'])
                    tmp_grant_agency.append([j.text for j in grant if j.tag=='Agency'])
            else:
                tmp_grant_ids.append(np.nan)
                tmp_grant_agency.append(np.nan)

            paper_level_meshlabs.append(tmp_mesh_labs)
            paper_level_meshids.append(tmp_mesh_ids)
            paper_level_grant_ids.append(tmp_grant_ids)
            paper_level_grant_agency.append(tmp_grant_agency)
            paper_level_kw.append(tmp_kw)
            pmid_out.append(paper[0][0].text)
    
        out_df = pd.DataFrame({'mesh':paper_level_meshlabs,
                               'mesh_id':paper_level_meshids,
                               'grant_ids':paper_level_grant_ids,
                               'grant_agency':paper_level_grant_agency,
                               'keyword':paper_level_kw,
                               'pmid':pmid_out})
        time.sleep(.11)
            
    except Exception as e:
        print(str(i[0])+'*******'+str(i[199]))
        print(e)
        time.sleep(.11)

#Lots of these are just zero. missing data i guess
pd.Series(stragglers.pmid2==0).value_counts()

total_papes = total_papes + out_df.shape[0]
out_df.to_csv('/home/riddleta/ac_knowl/output/grant_identifiers/'+fn+str(total_papes)+'.csv', index=False)

li = []
all_files = glob.glob('/home/riddleta/ac_knowl/output/grant_identifiers/*')
for filename in all_files:
    tempdat = pd.read_csv(filename, index_col=None, header=0)
    li.append(tempdat)

done_dat = pd.concat(li, axis=0, ignore_index=True)

# reformat grant-pmid pairing
Now, I want a file that has a grant id, the grant agency, and pmid

grant_funded = done_dat[done_dat.grant_ids!='[nan]']
grant_funded.shape

grant_funded.head()

grant_funded['grant_ids_cleaned'] = grant_funded.grant_ids.apply(lambda x: re.sub("\[|'", '', x))
grant_funded['grant_agency_cleaned'] = grant_funded.grant_agency.apply(lambda x: re.sub("\[|'", '', x))
#re.sub("\[|\]|'", '', grant_funded.grant_ids.iloc[0]).split(',')

grant_funded.head()

def interleave_lists(l1, l2):
    if len(l1)>1:
        outlist = ["{}---{}".format(b_, a_) for a_, b_ in zip(l1, l2)]
    else:
        outlist = [l1[0]+'---'+l2[0]]
        
    return(outlist)

grant_funded['grant_list'] = grant_funded.apply(lambda x: 
                                                interleave_lists(x['grant_ids_cleaned'].split('], '), 
                                                                 x['grant_agency_cleaned'].split('], ')), axis=1)

grant_pmids = grant_funded.explode('grant_list')
grant_pmids = grant_pmids[['pmid', 'grant_list']]

grant_pmids['grant_list'] = grant_pmids.grant_list.apply(lambda x: re.sub('\]', '', x))

grant_pmids[['agency', 'number', 'x1', 'x2']] = grant_pmids.grant_list.str.split('---', expand=True)

#nothing in x1 or x2 are from the NIH
grant_pmids = grant_pmids[grant_pmids.x1.isnull()]

grant_pmids.to_csv('output/pmcidextracted_grant_pmid_pairing.csv', index=False)

agency_counts = grant_pmids.agency.value_counts()
grant_id_counts = grant_pmids.number.value_counts()

agency_counts.iloc[0:10]

agency_counts.iloc[10:25]

agency_counts.iloc[25:50]

75*100

7500/175
