==> 4_Pull hit contexts.ipynb <==
import pandas as pd
import glob
import json
import re
from urlextract import URLExtract
pd.set_option('display.max_colwidth', -1)

# Pulling the context of a hit

This notebook contains code to go through the aggregated data, pulling out the context of a hit. This will facilitate checking how often our regex patterns are leading to false positives

extractor = URLExtract()

def hit_contexts(r_out):
    out_dat = []
    
    reg_matches = re.compile(r"""(github)|(osf\.io)|(nda\.nih\.gov)|(openneuro)|(\sndar)|
                                 (national database for autism research)|(brain-map\.org)|
                                 (humanconnectome\.org)|(balsa\.wustl\.edu)|
                                 (loni\.usc\.edu)|(ida\.loni\.usc\.edu)|(fmridc)|(ccrns)|
                                 (datalad)|(dataverse)|(dbgap)|(nih\.gov\/gap)|(dryad)|
                                 (figshare)|(fcon_1000\.projects)|(nitrc)|
                                 (mcgill\.ca\/bic\/resources\/omega)|(xnat\.org)|(zenodo)|
                                 (opendata\.aws)|(arrayexpress)|(genbank)|
                                 (protein databank)|(\sgeo\s)|
                                 (stanford microarray database)""", re.X)
    

    try:
        tmp_doi = r_out['documents'][0]['passages'][0]['infons']['article-id_doi']
    except:
        tmp_doi = None
    try:
        tmp_pmcid = r_out['documents'][0]['id']
    except:
        tmp_pmcid = None
        
    for passage in r_out['documents'][0]['passages']:
        m = re.finditer(reg_matches, passage['text'].lower())
        if m:
            segments = [(max(0, item.start()-175),
                         min(len(passage['text']), item.end(0)+125),
                         item.group()) for item in m]
            
            try:
                section_type = passage['infons']['section_type']
            except:
                section_type=None
            
            for seg in segments:
                out_dat.append([passage['text'], 
                                seg[2], 
                                passage['offset'],
                                tmp_pmcid, 
                                tmp_doi, 
                                section_type])
                
        if extractor.has_urls(passage['text'].lower()):
            try:
                section_type = passage['infons']['section_type']
            except:
                section_type=None
                
            out_dat.append([passage['text'],
                            'url_hit',
                            passage['offset'],
                            tmp_pmcid,
                            tmp_doi,
                            section_type])
        else:
            out_dat.append([None, None, tmp_pmcid, tmp_doi, None])
   
        
    return(out_dat)

already_retrieved = glob.glob('/home/riddleta/ac_knowl/output/full_texts/papes*')
already_retrieved = sorted(already_retrieved)
status_prints = already_retrieved[::3]
len(status_prints)

data_collect = []
for i in already_retrieved:
    if i in status_prints:
        print(i)
    with open(i) as infile:
        dat = json.load(infile)
        for j, paper in enumerate(dat):
            out_dat = hit_contexts(paper)
            data_collect.extend(out_dat)
              

data_collect = []
with open(already_retrieved[1]) as infile:
    dat = json.load(infile)
    for j, paper in enumerate(dat):
        out_dat = hit_contexts(paper)
        data_collect.extend(out_dat)

len(data_collect)

#df = pd.DataFrame(data_collect)
#df.columns = ['context', 'repo_hit', 'paper_offset', 'pmcid', 'doi', 'section']
df['pmcid_offset'] = df.pmcid.astype(str) + df.paper_offset.astype(str)
df.drop_duplicates('pmcid_offset', inplace=True)
df = df[~pd.isna(df.repo_hit)]
df = df[['context', 'repo_hit', 'paper_offset', 'pmcid', 'doi', 'section']]
df.to_csv('/home/riddleta/ac_knowl/output/hit_contexts.csv', index=False)

#github - all kinds of shit
#dryad - mostly posting data!
#figshare - mostly posting data!
#dbgap - mixture of posting data and reuse
#zenodo - mixture? zenodo seems to host lots of things
#nitrc - too general of a search term? includes lots o refs to BrainNet Viewer
#osf - mostly data!
#brain-map.org - mostly refs to already existing data
#loni - refs to existing data
#dataverse - mixture. seems like there are multiple things with the title 'dataverse'
#ndar - not many direct links. Lots of discussion. Plus a couple of things that aren't ndar: (nDart1 (gene?), tGuet Ndar (?), numpy ndarray...)
#nih.gov/gap - lots of reuse
#fcon_1000 - looks like lots of reuse
#humanconnectome - lots of reuse
#fmridc - seems mostly people talking about fmridc
#openneuro - looks like data posting
#ccrns seems like an acronym for something else
#ida.loni.usc.edu - downloads
#balsa - data posting!
#xnat - some data sharing? lots of things that don't appear to be data
#datalad - posting data and talking about datalad
#opendata.aws - one instance of reuse
#omega - one instance of just a link (appears to be reuse)

# % of 2.6 mil that are nimh funded
# percentage of all funded nimh that is in the 2.6 million https://federalreporter.nih.gov/FileDownload
#other tools to see if we can get a sense of all papers from the same timeframe.

sample_hits = df[df.repo_hit=='zenodo']
sample_hits.sample(n=5)

with open('/home/riddleta/ac_knowl/output/full_texts/papes_1402500.txt') as infile:
        dat = json.load(infile)

dat[2439]#['documents'][0]['passages'][17]
