==> 6_create_data_for_labeling.ipynb <==
import pandas as pd
import re
import glob
import json
from bs4 import BeautifulSoup
import random
import numpy as np
from urlextract import URLExtract
import spacy
from nltk import tokenize
import nltk.data
from nltk.tokenize.punkt import PunktSentenceTokenizer
import pickle

# Making datasets for labeling
There are several stages to this, and I'm going to pull samples to label from three sources.

First, I'm limiting this to the subset of papers that are funded by the NIMH.

Using that subset, I'm going to label 'passages' for whether they contain an instance of data sharing. A passage roughly corresponds to a paragraph in the paper, but sometimes is a footnote or a reference or a title. It's an organizational unit within the PMC fulltext databnase.

The sources that I'm pulling these passages from are:

1. Contexts previously identified as possible data-sharing using regular expressions.
2. A list of papers described as having shared data using NDAR. For this, I'll attempt to use any passages where data sharing is described
3. A random selection of passsages that are unlikely to contain data-sharing.

### Update
We've decided to re-label at the sentence level. After the above, toward the end of this document collects all the labels and makes a new sentence-level file.

#function to pull segments for labeling
extractor = URLExtract()

def hit_contexts(r_out, include_urls = False):
    out_dat = []
    
    reg_matches = re.compile(r"""(github)|(osf\.io)|(nda\.nih\.gov)|(openneuro)|(\sndar)|
                                 (national database for autism research)|(brain-map\.org)|
                                 (humanconnectome\.org)|(balsa\.wustl\.edu)|(loni\.usc\.edu)|
                                 (ida\.loni\.usc\.edu)|(fmridc)|(ccrns)|(datalad)|(dataverse)|
                                 (dbgap)|(nih\.gov\/gap)|(dryad)|(figshare)|(fcon_1000\.projects)|
                                 (nitrc)|(mcgill\.ca\/bic\/resources\/omega)|(xnat\.org)|
                                 (zenodo)|(opendata\.aws)""", re.X)
    

    try:
        tmp_doi = r_out['documents'][0]['passages'][0]['infons']['article-id_doi']
    except:
        tmp_doi = None
    try:
        tmp_pmcid = r_out['documents'][0]['id']
    except:
        tmp_pmcid = None
        
    for passage in r_out['documents'][0]['passages']:
        passage_marked = 0
        m = re.finditer(reg_matches, passage['text'].lower())
        if m:
            segments = [(max(0, item.start()-175),
                         min(len(passage['text']), item.end(0)+125),
                         item.group()) for item in m]
            
            try:
                section_type = passage['infons']['section_type']
            except:
                section_type=None
            
            for seg in segments:
                out_dat.append([passage['text'], #delete the seg[0]:seg[1] and you'll have the full segment
                                seg[2], #ids which repo
                                passage['offset'], #how far into the paper?
                                tmp_pmcid, 
                                tmp_doi, 
                                section_type])
            passage_marked = 1
        
        if include_urls:
            if extractor.has_urls(passage['text'].lower()):
                try:
                    section_type = passage['infons']['section_type']
                except:
                    section_type=None

                out_dat.append([passage['text'],
                                'url_hit',
                                passage['offset'],
                                tmp_pmcid,
                                tmp_doi,
                                section_type])
                passage_marked = 1
            
        if passage_marked == 0:
            out_dat.append([None, None, None, tmp_pmcid, tmp_doi, None])
        
    return(out_dat)

def sentence_hit(text, include_urls = False):
    passage_marked = 0
    
    reg_matches = re.compile(r"""(github)|(osf\.io)|(nda\.nih\.gov)|(openneuro)|(\sndar)|
                                 (national database for autism research)|(brain-map\.org)|
                                 (humanconnectome\.org)|(balsa\.wustl\.edu)|(loni\.usc\.edu)|
                                 (ida\.loni\.usc\.edu)|(fmridc)|(ccrns)|(datalad)|(dataverse)|
                                 (dbgap)|(nih\.gov\/gap)|(dryad)|(figshare)|(fcon_1000\.projects)|
                                 (nitrc)|(mcgill\.ca\/bic\/resources\/omega)|(xnat\.org)|
                                 (zenodo)|(opendata\.aws)""", re.X)
    
    m = re.findall(reg_matches, text.lower())
    if m:
        passage_marked = 1
        
    if include_urls:
        if extractor.has_urls(text.lower()):
            passage_marked = 1
        
    return(passage_marked)

def sample_section(paper):
    candidate = []
    for i in paper['documents'][0]['passages']:
        if i['infons']['section_type'] != 'REF':
            candidate.append(i)
    
    selection = random.choice(candidate)
    
    return(selection)

def test_suitability(passage):
    #figures out whether a passage is suitable for what we're after.
    #probably still needs some fiddling.
    suitability = True
    try:
        if passage['infons']['section_type'] == 'REF':
            suitability = False

        if passage['infons']['type'] == 'table':
            suitability = False

        if 'title' in passage['infons']['type']:
            suitability = False
    except:
        suitability = False
    
    return(suitability)

## Contexts previously identified as probable data sharing (via regex)

# load previously identified hits and papers funded by nimh
hits = pd.read_csv('output/hit_contexts.csv') #need the subset that is in nimh funded papers
nimh_papers = pd.read_csv('output/nimh_papers.csv')
#load file index
file_ix = pd.read_csv('output/file_index.csv')

#filter the hits to just nimh papers
nimh_hits = hits[hits.pmcid.isin(nimh_papers.pmcid)]
nimh_hits.shape

#pull a sample of 100 hits
sampled_papers = pd.Series(nimh_hits.pmcid.unique()).sample(1000, replace=False)
sampled_papers.shape

#get location of sampled hits
file_locs = file_ix[file_ix.pmcid.isin(sampled_papers)]

file_locs.shape

data_collect = []
for i, file_path in enumerate(file_locs.file):
    with open(file_path[24:]) as infile:
        dat = json.load(infile)
        paper = dat[file_locs.paper_number.iloc[i]]
        out = hit_contexts(paper)
        data_collect.extend(out)

df_to_label = pd.DataFrame(data_collect, columns=['text', 'repo', 'paper_offset', 'pmcid', 'doi', 'section'])
df_to_label = df_to_label.drop_duplicates(subset=['pmcid'])
df_to_label.to_csv('output/labeled_data/regex_hits.csv', index=False)

## Papers identified through NDAR

This is a convoluted process. First, I pull out all the papers mentioned in scraped NDA collections as being relevant. Next, I got a sample of those papers, and manually labeled them for containing instances of data sharing. With those labeled papers, I pulled out the section of text in each paper that mentions data sharing. If there was no such mention, then I just pulled out a random section (leaving out anything that was a reference). I then put these data together and wrote them as a csv file (`ndar_labs.csv`). Though I aimed for a sample of 100, I ended up with 65 due to a few papers being embargoed, and some papers listed in NDAR not being in the full-text database that I have.

ndar_collections = glob.glob('output/ndar_collections/*')
ndar_collections.sort()

len(ndar_collections)

l = []
for collection in ndar_collections:
    soup = BeautifulSoup(open(collection), "html.parser")
    table = soup.find_all(id='publication-table')[0]
    table_rows = table.find_all('tr')
    for tr in table_rows:
        td = tr.find_all('td')
        row = [tr.text for tr in td]
        l.append(row)
        
collection_pubs = pd.DataFrame(l, columns = ['pmid', 'study', 'title', 'journal', 'authors', 'date', 'status'])

pd.unique(collection_pubs.study).shape

#pick out the pmids for the studies marked as relevant
collection_pmids = pd.Series(collection_pubs.pmid[collection_pubs.status=='Relevant'].unique())
len(collection_pmids)

#read in the linking file and convert pmids to integer
pmid_to_pmcid = pd.read_csv('data/PMC-ids.csv')
pmid_to_pmcid['pmid'] = pmid_to_pmcid.PMID.fillna(0.0).astype(int)

pmid_to_pmcid.head()

#keep the rows in the linking file that have a match from ndar; strip out letters from pmcid
referenced_papers = pmid_to_pmcid[pmid_to_pmcid.pmid.isin(collection_pmids)]
referenced_papers['pmcid'] = referenced_papers.PMCID.apply(lambda x: x[3:])
print(referenced_papers.shape)
referenced_papers.Year.value_counts()

#pull a sample of 100 hits
sampled_papers = pd.Series(referenced_papers.pmcid.unique()).sample(100, replace=False)
#this file was manually labeled so that I can get specific passages
sampled_papers.to_csv('output/ndar_pmcids.csv', index=False) 

sampled_papers = pd.read_csv('output/ndar_pmcids.csv')
#split out those that have shared and not shared data
no_shares = sampled_papers[sampled_papers.data_sharing=='0']
shares = sampled_papers[sampled_papers.data_sharing=='1']

no_shares_locs = file_ix[file_ix.pmcid.isin(no_shares.PMCID.tolist())]
text = []
repo = []
paper_offset = []
pmcid = []
doi = []
section_type = []
for i, file_path in enumerate(no_shares_locs.file):
    with open(no_shares_locs.file.iloc[i][24:]) as infile:
        dat = json.load(infile)
        paper = dat[no_shares_locs.paper_number.iloc[i]]
        section = sample_section(paper)
        text.append(section['text'])
        repo.append(np.nan)
        paper_offset.append(section['offset'])
        pmcid.append(no_shares_locs.pmcid.iloc[i])
        doi.append(np.nan)
        section_type.append(section['infons']['section_type'])
        
dat = pd.DataFrame({'text':text,
                    'repo':repo,
                    'paper_offset':paper_offset,
                    'pmcid':pmcid,
                    'doi':doi,
                    'section':section_type})

shares_locs = file_ix[file_ix.pmcid.isin(shares.PMCID.tolist())]
text = []
repo = []
paper_offset = []
pmcid = []
doi = []
section_type = []


with open(shares_locs.file.iloc[0][24:]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[0]]
    text.append(paper['documents'][0]['passages'][77]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][77]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[0])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][77]['infons']['section_type'])

with open(shares_locs.file.iloc[2][24:]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[2]]
    text.append(paper['documents'][0]['passages'][62]['text'])
    repo.append('GEO')
    paper_offset.append(paper['documents'][0]['passages'][62]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[2])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][62]['infons']['section_type'])

with open(shares_locs.file.iloc[3][24:]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[3]]
    text.append(paper['documents'][0]['passages'][88]['text'])
    repo.append('OSF')
    paper_offset.append(paper['documents'][0]['passages'][88]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[3])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][88]['infons']['section_type'])

with open(shares_locs.file.iloc[4][24:]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[4]]
    text.append(paper['documents'][0]['passages'][47]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][47]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[4])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][47]['infons']['section_type'])

with open(shares_locs.file.iloc[5][24:]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[5]]
    text.append(paper['documents'][0]['passages'][91]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][91]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[5])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][91]['infons']['section_type'])

dat2 = pd.DataFrame({'text':text,
                    'repo':repo,
                    'paper_offset':paper_offset,
                    'pmcid':pmcid,
                    'doi':doi,
                    'section':section_type})

ndar_labs = pd.concat([dat, dat2])

ndar_labs.to_csv('output/labeled_data/ndar_labs.csv', index=False)

## Now do a random selection of other passages

nimh_papers = pd.read_csv('output/nimh_papers.csv')
#load file index
file_ix = pd.read_csv('output/file_index.csv')

#pull a sample of 100 hits
sampled_papers = pd.Series(nimh_papers.pmcid.unique()).sample(100, replace=False)
sampled_papers.shape

file_locs = file_ix[file_ix.pmcid.isin(sampled_papers)]
file_locs.shape

out_dat = []
for i, file_path in enumerate(file_locs.file):
    with open(file_path[24:]) as infile:
        dat = json.load(infile)
        paper = dat[file_locs.paper_number.iloc[i]]
        sec = sample_section(paper)
        out_dat.append([sec['text'], #delete the seg[0]:seg[1] and you'll have the full segment
                np.nan, #ids which repo
                sec['offset'], #how far into the paper?
                file_locs.pmcid.iloc[i], 
                np.nan, 
                sec['infons']['section_type']])
        


temp = pd.DataFrame(out_dat, columns=['text', 'repo', 'paper_offset', 'pmcid', 'doi', 'section'])
temp.to_csv('output/labeled_data/random_selections.csv', index=False)

# We need to add some additional cases

First, I'm going to pull out some of the instances identified via regex matches.

# load previously identified hits and papers funded by nimh
hits = pd.read_csv('output/hit_contexts.csv') #need the subset that is in nimh funded papers
nimh_papers = pd.read_csv('output/nimh_papers.csv')
#load file index
file_ix = pd.read_csv('output/file_index.csv')
#filter the hits to just nimh papers
nimh_hits = hits[hits.pmcid.isin(nimh_papers.pmcid)]
nimh_hits.shape

# filter out the previously sampled cases
excludes = pd.read_csv('output/labeled_data/regex_hits.csv')
nimh_hits = nimh_hits[~nimh_hits.pmcid.isin(excludes.pmcid)]
print(nimh_hits.shape)
#pull a sample of 100 hits
sampled_papers = pd.Series(nimh_hits.pmcid.unique()).sample(100, replace=False)
sampled_papers.shape

#get location of sampled hits
file_locs = file_ix[file_ix.pmcid.isin(sampled_papers)]
data_collect = []
for i, file_path in enumerate(file_locs.file):
    with open(file_path[24:]) as infile:
        dat = json.load(infile)
        paper = dat[file_locs.paper_number.iloc[i]]
        out = hit_contexts(paper)
        data_collect.extend(out)

df_to_label = pd.DataFrame(data_collect, columns=['text', 'repo', 'paper_offset', 'pmcid', 'doi', 'section'])
df_to_label = df_to_label.drop_duplicates(subset=['pmcid'])
df_to_label.to_csv('output/labeled_data/regex_hits2.csv', index=False)

## Additional cases - NDAR selections2

ndar_collections = glob.glob('output/ndar_collections/*')
ndar_collections.sort()

l = []
for collection in ndar_collections:
    soup = BeautifulSoup(open(collection), "html.parser")
    table = soup.find_all(id='publication-table')[0]
    table_rows = table.find_all('tr')
    for tr in table_rows:
        td = tr.find_all('td')
        row = [tr.text for tr in td]
        l.append(row)
        
collection_pubs = pd.DataFrame(l, columns = ['pmid', 'study', 'title', 'journal', 'authors', 'date', 'status'])

#pick out the pmids for the studies marked as relevant
collection_pmids = pd.Series(collection_pubs.pmid[collection_pubs.status=='Relevant'].unique())

#read in the linking file and convert pmids to integer
pmid_to_pmcid = pd.read_csv('data/PMC-ids.csv')
pmid_to_pmcid['pmid'] = pmid_to_pmcid.PMID.fillna(0.0).astype(int)

#keep the rows in the linking file that have a match from ndar; strip out letters from pmcid
referenced_papers = pmid_to_pmcid[pmid_to_pmcid.pmid.isin(collection_pmids)]
referenced_papers['pmcid'] = referenced_papers.PMCID.apply(lambda x: x[3:])
 

# filter to just nimh funded work
nimh_papers = pd.read_csv('output/nimh_papers.csv')
nimh_pmcids = nimh_papers.pmcid.astype('str').tolist()

referenced_papers = referenced_papers[referenced_papers.pmcid.isin(nimh_pmcids)]

#filter out the already-labeled ndar data
df_ndar_labs = pd.read_csv('output/labeled_data/ndar_labs.csv')
labeled_ndars = df_ndar_labs.pmcid.astype('str').tolist()
referenced_papers = referenced_papers[~referenced_papers.pmcid.isin(labeled_ndars)]

#get the file location for the matches
#pull a sample of 100 hits
sampled_papers = pd.Series(referenced_papers.pmcid.unique()).sample(100, replace=False)

#this file was manually labeled so that I can get specific passages
sampled_papers.to_csv('output/ndar_pmcids2.csv', index=False)

sampled_papers = pd.read_csv('output/ndar_pmcids2.csv')
file_ix = pd.read_csv('output/file_index.csv')
#split out those that have shared and not shared data
no_shares = sampled_papers[sampled_papers.open_data==0]
shares = sampled_papers[sampled_papers.open_data==1]

no_shares_locs = file_ix[file_ix.pmcid.astype('str').isin(no_shares.PMCID.astype('str').tolist())]
text = []
repo = []
paper_offset = []
pmcid = []
doi = []
section_type = []
for i, file_path in enumerate(no_shares_locs.file):
    with open(no_shares_locs.file.iloc[i][24:]) as infile:
        dat = json.load(infile)
        paper = dat[no_shares_locs.paper_number.iloc[i]]
        section = sample_section(paper)
        text.append(section['text'])
        repo.append(np.nan)
        paper_offset.append(section['offset'])
        pmcid.append(no_shares_locs.pmcid.iloc[i])
        doi.append(np.nan)
        section_type.append(section['infons']['section_type'])
        
dat = pd.DataFrame({'text':text,
                    'repo':repo,
                    'paper_offset':paper_offset,
                    'pmcid':pmcid,
                    'doi':doi,
                    'section':section_type})

shares_locs = file_ix[file_ix.pmcid.astype('str').isin(shares.PMCID.astype('str').tolist())]
text = []
repo = []
paper_offset = []
pmcid = []
doi = []
section_type = []

i = 0
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][61]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][61]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][61]['infons']['section_type'])

i = 1
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][16]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][16]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][16]['infons']['section_type'])

i = 2
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][50]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][50]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][50]['infons']['section_type'])

i = 3
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][15]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][15]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][15]['infons']['section_type'])

i = 4
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][42]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][42]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][42]['infons']['section_type'])

i = 5
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][11]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][11]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][11]['infons']['section_type'])

i = 6
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][49]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][49]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][49]['infons']['section_type'])

i = 7
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][83]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][83]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][83]['infons']['section_type'])

i = 8
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][26]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][26]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][26]['infons']['section_type'])

i = 9
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][43]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][43]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][43]['infons']['section_type'])

i = 10
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][11]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][11]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][11]['infons']['section_type'])

i = 11
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][44]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][44]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][44]['infons']['section_type'])

i = 12
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][60]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][60]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][60]['infons']['section_type'])

i = 13
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][54]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][54]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][54]['infons']['section_type'])

i = 14
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][45]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][45]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][45]['infons']['section_type'])

i = 15
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][11]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][11]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][11]['infons']['section_type'])

i = 16
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][54]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][54]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][54]['infons']['section_type'])

i = 17
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][51]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][51]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][51]['infons']['section_type'])

dat2 = pd.DataFrame({'text':text,
                    'repo':repo,
                    'paper_offset':paper_offset,
                    'pmcid':pmcid,
                    'doi':doi,
                    'section':section_type})

ndar_labs = pd.concat([dat, dat2])

sampled_papers['pmcid'] = sampled_papers.PMCID.astype('str')
ndar_labs['pmcid'] = ndar_labs.pmcid.astype('str')
ndar_labs = ndar_labs.merge(sampled_papers[['pmcid', 'data_sharing', 'open_data']], how='left', on='pmcid')

ndar_labs.to_csv('output/labeled_data/ndar_labs2.csv', index=False)

# all nimh passages

nimh_papers = pd.read_csv('output/nimh_papers.csv')
#load file index
file_ix = pd.read_csv('output/file_index.csv')

file_ix['pmcid'] = file_ix.pmcid.astype('str')

nimh_papers['pmcid'] = nimh_papers.pmcid.astype('str')

target_papers = file_ix[file_ix.pmcid.isin(nimh_papers.pmcid)]
target_papers.shape

target_papers = target_papers.sort_values('file')

target_papers.head()

status_prints = range(0, len(target_papers.file.tolist()), 250)
len(status_prints)

data_collect = []
last_file = np.nan
for i, file in enumerate(target_papers.file.tolist()):
    if i in status_prints:
        print(i)
    if file == last_file:
        paper = dat[target_papers.paper_number.iloc[i]]
        out_dat = hit_contexts(paper, include_urls=True)
        data_collect.extend(out_dat)
    else:
        with open(file) as infile:
            dat = json.load(infile)
            paper = dat[target_papers.paper_number.iloc[i]]
            out_dat = hit_contexts(paper, include_urls=True)
            data_collect.extend(out_dat)
            last_file = file

len(data_collect)

df = pd.DataFrame(data_collect)
df.columns = ['context', 'repo_hit', 'paper_offset', 'pmcid', 'doi', 'section']
df.head()

df.to_csv('output/hit_contexts_nimh_v2.csv', index=False)

out_dat['pmcid_url'] = out_dat.pmcid.apply(lambda x: 'http://ncbi.nlm.nih.gov/pmc/articles/PMC'+str(x))
out_dat['doi_url'] = out_dat.doi.apply(lambda x: 'http://doi.org/'+str(x))
out_dat.to_excel('output/hit_contexts.nimh.xlsx', index=False)

len(out_dat.pmcid.value_counts())

## Pull pubs out of HCP

soup = BeautifulSoup(open('data/HCP_Publications.htm'), "html.parser")
pubs = soup.find_all('div', {"class": "publication-data-wrapper"})

l = []
for i in pubs:
    title = i.find('h4').text.strip()
    auths = i.find('div', {'class':'publication-authors'}).text.strip()
    journal = i.find('span', {'class':'publication-data-name'}).text.strip()
    date = i.find('span', {'class':'publication-data-date'}).text.strip()
    try:
        pmid = i.find('span', {'class':'publication-data-pmid'}).text.strip()[6:]
    except:
        pmid = np.nan
    try:
        doi_href = i.find('span', {'class':'publication-data-name'}).find('a')['href']
    except:
        doi_href = np.nan
    l.append([title, auths, journal, date, pmid, doi_href])

df = pd.DataFrame(l, columns = ['title', 'auths', 'journal', 'date', 'pmid', 'doi_href'])

df.to_csv('output/hcp_pubs.csv', index=False)

## Pull out some passages identified via regex (again)

# load previously identified hits and papers funded by nimh
hits = pd.read_csv('output/hit_contexts.csv') #need the subset that is in nimh funded papers
nimh_papers = pd.read_csv('output/nimh_papers.csv')
#load file index
file_ix = pd.read_csv('output/file_index.csv')
#filter the hits to just nimh papers
nimh_hits = hits[hits.pmcid.isin(nimh_papers.pmcid)]
nimh_hits.shape

# filter out the previously sampled cases
excludes = pd.read_csv('output/labeled_data/regex_hits.csv')
nimh_hits = nimh_hits[~nimh_hits.pmcid.isin(excludes.pmcid)]
print(nimh_hits.shape)
#pull a sample of 100 hits
sampled_papers = pd.Series(nimh_hits.pmcid.unique()).sample(100, replace=False)
sampled_papers.shape

#get location of sampled hits
file_locs = file_ix[file_ix.pmcid.isin(sampled_papers)]
data_collect = []
for i, file_path in enumerate(file_locs.file):
    with open(file_path[24:]) as infile:
        dat = json.load(infile)
        paper = dat[file_locs.paper_number.iloc[i]]
        out = hit_contexts(paper)
        data_collect.extend(out)

df_to_label = pd.DataFrame(data_collect, columns=['text', 'repo', 'paper_offset', 'pmcid', 'doi', 'section'])
df_to_label = df_to_label.drop_duplicates(subset=['pmcid'])
#df_to_label.to_csv('output/labeled_data/regex_hits3.csv', index=False)
df_to_label.to_csv('output/labeled_data/regex_hits4.csv', index=False)

df_to_label.head()

## Additional cases - NDAR selections3

ndar_collections = glob.glob('output/ndar_collections/*')
ndar_collections.sort()

l = []
for collection in ndar_collections:
    soup = BeautifulSoup(open(collection), "html.parser")
    table = soup.find_all(id='publication-table')[0]
    table_rows = table.find_all('tr')
    for tr in table_rows:
        td = tr.find_all('td')
        row = [tr.text for tr in td]
        l.append(row)
        
collection_pubs = pd.DataFrame(l, columns = ['pmid', 'study', 'title', 'journal', 'authors', 'date', 'status'])

#pick out the pmids for the studies marked as relevant
collection_pmids = pd.Series(collection_pubs.pmid[collection_pubs.status=='Relevant'].unique())

#read in the linking file and convert pmids to integer
pmid_to_pmcid = pd.read_csv('data/PMC-ids.csv')
pmid_to_pmcid['pmid'] = pmid_to_pmcid.PMID.fillna(0.0).astype(int)

#keep the rows in the linking file that have a match from ndar; strip out letters from pmcid
referenced_papers = pmid_to_pmcid[pmid_to_pmcid.pmid.isin(collection_pmids)]
referenced_papers['pmcid'] = referenced_papers.PMCID.apply(lambda x: x[3:])
 

# filter to just nimh funded work
nimh_papers = pd.read_csv('output/nimh_papers.csv')
nimh_pmcids = nimh_papers.pmcid.astype('str').tolist()

referenced_papers = referenced_papers[referenced_papers.pmcid.isin(nimh_pmcids)]

#filter out the already-labeled ndar data
ndar_lab_files = glob.glob('output/labeled_data/ndar*.csv')
df_ndar_labs = pd.concat((pd.read_csv(f) for f in ndar_lab_files)).reset_index()
labeled_ndars = df_ndar_labs.pmcid.astype('str').tolist()
referenced_papers = referenced_papers[~referenced_papers.pmcid.isin(labeled_ndars)]

#get the file location for the matches
#pull a sample of 100 hits
sampled_papers = pd.Series(referenced_papers.pmcid.unique()).sample(100, replace=False)

#this file was manually labeled so that I can get specific passages
sampled_papers.to_csv('output/ndar_pmcids3.csv', index=False)

sampled_papers = pd.read_csv('output/ndar_pmcids3.csv')
sampled_papers['pmcid'] = sampled_papers.PMCID.astype('int').astype('str')
file_ix = pd.read_csv('output/file_index.csv')
#split out those that have shared and not shared data
no_shares = sampled_papers[sampled_papers.open_data==0]
shares = sampled_papers[sampled_papers.open_data==1]

no_shares_locs = file_ix[file_ix.pmcid.astype('str').isin(no_shares.PMCID.astype('int').astype('str').tolist())]
text = []
repo = []
paper_offset = []
pmcid = []
doi = []
section_type = []
for i, file_path in enumerate(no_shares_locs.file):
    with open(no_shares_locs.file.iloc[i][24:]) as infile:
        dat = json.load(infile)
        paper = dat[no_shares_locs.paper_number.iloc[i]]
        section = sample_section(paper)
        text.append(section['text'])
        repo.append(np.nan)
        paper_offset.append(section['offset'])
        pmcid.append(no_shares_locs.pmcid.iloc[i])
        doi.append(np.nan)
        section_type.append(section['infons']['section_type'])
        
dat = pd.DataFrame({'text':text,
                    'repo':repo,
                    'paper_offset':paper_offset,
                    'pmcid':pmcid,
                    'doi':doi,
                    'section':section_type})

shares_locs = file_ix[file_ix.pmcid.astype('str').isin(shares.PMCID.astype('int').astype('str').tolist())]
text = []
repo = []
paper_offset = []
pmcid = []
doi = []
section_type = []

shares_locs.shape

i = 0
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][31]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][31]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][31]['infons']['section_type'])
i = 1
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][14]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][14]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][14]['infons']['section_type'])
i = 2
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][0]['infons']['notes'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][0]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append('NOTES')
i = 3
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][26]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][26]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][26]['infons']['section_type'])
i = 4
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][40]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][40]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][40]['infons']['section_type'])
i = 5
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][19]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][19]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][19]['infons']['section_type'])
i = 6
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][15]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][15]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][15]['infons']['section_type'])
i = 7
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][47]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][47]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][47]['infons']['section_type'])
i = 8
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][77]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][77]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][77]['infons']['section_type'])
i = 9
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][34]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][34]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][34]['infons']['section_type'])
i = 10
with open(shares_locs.file.iloc[i]) as infile:
    share_dat = json.load(infile)
    paper = share_dat[shares_locs.paper_number.iloc[i]]
    text.append(paper['documents'][0]['passages'][84]['text'])
    repo.append('NDAR')
    paper_offset.append(paper['documents'][0]['passages'][84]['offset'])
    pmcid.append(shares_locs.pmcid.iloc[i])
    doi.append(np.nan)
    section_type.append(paper['documents'][0]['passages'][84]['infons']['section_type'])

dat2 = pd.DataFrame({'text':text,
                    'repo':repo,
                    'paper_offset':paper_offset,
                    'pmcid':pmcid,
                    'doi':doi,
                    'section':section_type})

ndar_labs = pd.concat([dat, dat2])
ndar_labs['pmcid'] = ndar_labs.pmcid.astype('str')
ndar_labs = ndar_labs.merge(sampled_papers[['pmcid', 'data_sharing', 'open_data']], how='left', on='pmcid')
ndar_labs.to_csv('output/labeled_data/ndar_labs3.csv', index=False)

## Convert to sentence level

It turns out that out-of-the-box sentence tokenizers don't work too well on these scientific texts. I experimented with the punkt tokenizer in nltk and it looks like it works pretty well, but for whatever reason, it's pretty sensitive to the type of text it is given. I'm training it on the data that is marked as shared, as it yields good performance. I suspect there might be some really mal-formed text in the larger sample and may need to filter out tables, figures, etc. If I'm right about this, then the current approach is just a quick-and-dirty way of accomplishing that.

lab_files = glob.glob('output/labeled_data/*.csv')
df_labs = pd.concat((pd.read_csv(f) for f in lab_files)).reset_index()
al_files_ds = glob.glob('output/labeled_data/active_learning_labels/data*.csv')
al_files_os = glob.glob('output/labeled_data/active_learning_labels/open*.csv')
df_al_ds = pd.concat((pd.read_csv(f) for f in al_files_ds)).reset_index()
df_al_os = pd.concat((pd.read_csv(f) for f in al_files_os)).reset_index()

tmp = pd.concat([df_labs, df_al_ds, df_al_os]).reset_index(drop=True)
df_ds = tmp[tmp.data_sharing==1]
df_od = tmp[tmp.open_data==1]
tmp_text = pd.concat([df_ds, df_od])
tmp_text = tmp_text[['paper_offset', 'pmcid', 'section', 'text']].drop_duplicates()
train_text = tmp_text.text.tolist()
train_text = ' '.join(train_text)
tokes = PunktSentenceTokenizer(train_text)
tokes.tokenize('James et al. (1991) show evidence of this effect can be seen in Fig. 7 and in fig. 8. We next, along with dr. me and dr. doe try to show that mri is too expensive.')

out = open('output/tokenizer.pk', 'wb')
pickle.dump(tokenizer, out)
out.close()

tmp_text['text'] = tmp_text.text.apply(lambda x: tokes.tokenize(x))
tmp_text = tmp_text.explode('text')
tmp_text.to_csv('output/labeled_data/sentences.csv', index=False)

## Sentence level data for Adam

df_labeled = pd.read_csv('output/labeled_data/sentences.csv')
df_pmcids = pd.read_csv('data/PMC-ids.csv')
tk_file = open('output/tokenizer.pk', 'rb')
tokenizer = pickle.load(tk_file)
tk_file.close()

ds_files = glob.glob('output/labeled_data/active_learning_labels_sent/data_sharing*.csv')
od_files = glob.glob('output/labeled_data/active_learning_labels_sent/open_data*.csv')
df_data_sharing_al = pd.concat((pd.read_csv(f) for f in ds_files))
df_open_data_al = pd.concat((pd.read_csv(f) for f in od_files))

df = pd.concat([df_labeled, df_data_sharing_al, df_open_data_al])

df.text.fillna('', inplace=True)
df.reset_index(inplace=True, drop=True)
df.open_data[df.data_sharing==1] = 1
df.head()

df[df.open_data==1].to_csv('output/labeled_data/adam_sentences.csv', index=False)

df_contexts = pd.read_csv('output/hit_contexts.csv')
nimh_papers = pd.read_csv('output/nimh_papers.csv')
nimh_hits = df_contexts[df_contexts.pmcid.isin(nimh_papers.pmcid)]
nimh_hits.shape

nimh_hits.head()

nimh_hits['context'] = nimh_hits.context.apply(lambda x: tokenizer.tokenize(x))
nimh_hits = nimh_hits.explode('context')
nimh_hits['context'] = nimh_hits.context.fillna('')
nimh_hits.head()
#the below was using 'hits' detected at the section level
#sampled_hits = nimh_hits.sample(1000, random_state=42)
#sampled_hits.reset_index(drop=True).to_csv('output/labeled_data/adam_sentences_regex.csv', index=False)

nimh_hits['sentence_hit'] = nimh_hits.context.apply(lambda x: sentence_hit(x, include_urls=True))

nimh_hits.sentence_hit.value_counts()

nimh_sentence_hits = nimh_hits[nimh_hits.sentence_hit==1]
sampled_hits = nimh_sentence_hits.sample(1000, random_state=42)
sampled_hits.reset_index(drop=True).to_csv('output/labeled_data/adam_sentences_regex2.csv', index=False)
